+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Fixed Gauss gun dealing full damage when saving and loading right after starting a charged shot (Thanks Oxofemple.)
 * Prevent breakables from spawning multiple items when destroyed by gunfire and explosives at the same time (Thanks Oxofemple.)
 * Fixed save game system not saving arrays of EHANDLEs if the first half of the array contains null handles (mainly affected Nihilanth's spheres) [#224](https://github.com/SamVanheer/halflife-updated/issues/224) (Thanks Ronin4862)
+* Fixed player gaining health when drowning with god mode enabled and recovering health after surfacing (Thanks malortie)
 
 ### Features